Migrate `tests/unit/modules/test_state.py` to Pytest
